clean the code
